ramsete pseudo coded

+++ include/misc/distance_reset.h
@@ -9,8 +9,8 @@ enum class Direction {UP, DOWN, LEFT, RIGHT};
 class DistanceReset {
     private:
 
-        float left_offset;
-        float right_offset;
+        double left_offset;
+        double right_offset;
     
     public:
 
@@ -22,12 +22,12 @@ class DistanceReset {
          * @brief Distance Reset Constructor 
          * 
          */
-        DistanceReset(pros::Distance* left_sensor, float left_offset, pros::Distance* right_sensor, float right_offset);
+        DistanceReset(pros::Distance* left_sensor, double left_offset, pros::Distance* right_sensor, double right_offset);
 
 
         void update_position(Sensor sensor_selection, Direction direction);
 
-        float mm_to_in(float mm);
+        double mm_to_in(double mm);
 };
 
 extern DistanceReset dist_reset;
